chore: cleaning project and refactor for directories
