+++ .gitignore
@@ -1,0 +1,11 @@
+# Node artifact files
+node_modules/
+dist/
+dist.zip
+.env
+
+# Generated by MacOS
+.DS_Store
+
+# Log files
+*.log
