+++ backend/lambdas/generic/genericLambda.ts
@@ -1,5 +1,5 @@
 import { Handler} from 'aws-lambda';
-import GenericController from '../../src/controllers/generic.controller';
+import { GenericController } from '../../src/controllers/generic.controller';
 
 export const genericFunction: Handler = async (event) => {
     const genericController = await GenericController.create();
