+++ backend/package.json
@@ -12,7 +12,7 @@
     "aws-sdk": "^2.996.0",
     "axios": "^0.21.4",
     "bcrypt": "^5.0.1",
-    "class-validator": "^0.13.1",
+    "class-validator": "^0.13.2",
     "crypto": "^1.0.1",
     "dotenv": "^10.0.0",
     "jsonwebtoken": "^8.5.1",
