+++ backend/src/controllers/index.ts
@@ -1,0 +1,3 @@
+export { AuthorizerController } from './authorizer.controller';
+export { BaseController } from './base.controller';
+export { GenericController } from './generic.controller';
