+++ backend/src/errors/index.ts
@@ -1,0 +1,4 @@
+export { CustomError } from './CustomError';
+export { BadRequestError } from './BadRequestError';
+export { UnauthorizedError } from './UnauthorizedError';
+export { InternalServerError } from './InternalServerError';
