+++ backend/src/model/index.ts
@@ -1,0 +1,1 @@
+export { User } from './user.model';
