+++ backend/src/services/index.ts
@@ -1,0 +1,2 @@
+export { GenericService } from './generic.service';
+export { DbService } from './db.service';
